Rework the detector and tracker model hierarchy.
This is to try and fi the serialization/deserialization mess.
We need a specific class for each detector and tracker model,
acknowledging the fact that some models might be created for
detectors that might be not available (e.g. the StarDist detector
on a Fiji that does not have StarDist installed).