Handle model de/serialization in weird case.
It is possible to run the help with missing modules. For instance
we might miss the Cellpose detector. In that case we just inform the user.

But then the model will be saved stating that the module was missing.
If later the user returns having installed the missed model, we want
to reinstate a proper model that takes into account the newly installed
module.
Similarly, things might happen the other way around with a user
uninstalling a module, with a reference to it in the JSon file.

This commit helps handling both cases gracefully.