Extendable exporter plugin framework. GUI part.
Now exporters for the batcher are TrackMateMOdules that can be
added in a generic waty. They declare methods that can automatically
generate the run param GUI. It is properly wired for listeners and
serialized to JSon properly. (Probably breaks if an exiting json file
from a previous version is found.)
Now on to execution of the export.