Better way of checking if a parameter set was already tested.
Prior to this commit, we checked whether a parameter set was already
tested only AFTER detection. It is a pity because detection is often
the longest step in tracking. So we needed to wait for the detection
to be over to check whether we also did all the possible tracking
parameters.

Now we test for already done parameters immediately before doing
detection and perform detection only if at least one detection
parameter set + one tracking parameter set has not been tested.
Much faster and lighter.