Rework metric computation in the runners.
When used in the Helper:

- An abstract method does the computation and return the metrics.
- A method in the mother class calls this method and takes care of
saving the results to a CSV file.

This diminishes slightly duplicated code, allows lowering visibility
of some methods in MetricsRunner,  but will also allow using the
runners just to compute metrics oustide of the helper.

Also for SPT runner: the units is specified at construction. Nevermind
if we cannot get it right, it is just for display.