Don't crash with SPT metrics and cell division events.
It's possible to make a tracking configuration that will generate
a branching event in a track. With this we will have two spots
in the same timepoint and in the same track. This is ok for the CTC
metrics but not for the SPT metrics, that trigger an exception.
With this commit we gracefully catch it and write that the tracking
parameters generate an invalid tracking results for SPT. And continue
the sweep.