Support comparing with and without spot and track filters.
Spot filters and track filters are stored in the results tables if
present. They are used to compare settings against others, and are
properly reported in the results table.
This has become especially important in the light of the new SPT
metrics integration, that has a metric for penalizing false positive.