Try to work around deserialization issues with Gson.
All deserialized models have listeners that need to be initialized
and registered to the main ParameterSweepModel. But when deserializing
Gson, the constructor of the base class is not called, and the
listeners are not re-registered.

We try to work around this in a better way:
- The listeners field is initialized in the class definition, not in the
constructor.

But this does not work all the time either. For a reason that completely
escapes me, sometimes an instance will not have this field initialized.
Weird.

- A default visibility method is called by the Gson deserialization
logic to re-register the listeners to the main model.
- The param models have also an initialize() method that recreates
the listeners.
- Special case for CondaEnvParamSweepModel: the env list is cleared
after deserialization and replaced by the env list discovered at
runtime.
- Also: all RangeType enums have been renamed according to what is their
mother param type for clarity.