Recuperate from having serialized a missing module.
We may have serialized that the module was not available or
properly configured at the time of serialization. In that
case, the submodel is only one information parameter. Instead
of returning it, we try to instantiate a new model, so that
the user is presented something par with the current
configuration they have.