Do not privilege the TARGET_CHANNEL setting
Before this commit, the target channel was set in the main window
of the GUI, and used for all detectors and trackers. But some
detectors do not need it, or specify it another way. It makes no
sense to separate it from the other parameters. It adds clutter
and confusion.

Now the modules that require the TARGET_CHANNEL must specify it
as a parameter, that can be sweeped upon as well.